+++ apple-codesign/CHANGELOG.md
@@ -22,6 +22,9 @@ Released on ReleaseDate.
 * The `--code-signature-flags` argument of the `sign` command now correctly
   applies multiple values. Before, flags were set to the final specified
   value.
+* Added several trycmd based tests for testing CLI and signing behaviors.
+  The trycmd tests may download a prebuilt Rust coreutils binary from
+  github.com when executing on platforms with prebuilt binaries.
 * aws crates 0.53 -> 0.57.
 * bitflags 1.3 -> 2.0.
 * cryptographic-message-syntax 0.19 -> 0.25.